Feature: Add load model fp16

+++ translator/gradio.py
@@ -4,11 +4,12 @@ from transformers import AutoTokenizer, AutoModelForSeq2SeqLM
 import os
 from peft import PeftModel
 import re
+import torch
 
 
 tokenizer = AutoTokenizer.from_pretrained("google/mt5-large")
-model = AutoModelForSeq2SeqLM.from_pretrained("google/mt5-large")
-model = PeftModel.from_pretrained(model, "lora/checkpoint-55000").to("cuda")
+model = AutoModelForSeq2SeqLM.from_pretrained("google/mt5-large", torch_dtype=torch.float16)
+model = PeftModel.from_pretrained(model, "lora/checkpoint-55000", torch_dtype=torch.float16).to("cuda")
 
 def preprocess(input_text, prefix):
     return prefix + input_text + "|<END>|"
